refactor: Event system overhaul with custom types and thin canvas handlers
- Create custom MouseEvent/KeyEvent types with helper methods (offset, with_coords, is_shift, etc.)
- Add FinishedDrawing action for tool-to-canvas communication
- Simplify canvas event handlers: thin forwarding layer that delegates to tools
- Move tool-specific logic to tool implementations (TextTool handles keyboard input)
- Remove element.rs, consolidate all element logic in state/elements/
- Tools now return FinishedDrawing action; canvas handles tool switching based on tool_locked

+++ src/state/selection.rs
@@ -200,8 +200,8 @@ impl SelectionState {
     }
 
     pub fn finish_move_selection(&mut self, canvas: &mut CanvasState) {
-        let dx = self.move_offset.0;
-        let dy = self.move_offset.1;
+        let dx = self.move_offset.0 as i16;
+        let dy = self.move_offset.1 as i16;
 
         // Move all selected elements by offset
         for element_id in &self.selected_ids {
@@ -223,7 +223,7 @@ impl SelectionState {
 
         for element_id in &self.selected_ids {
             if let Some(element) = canvas.get_element_mut(*element_id) {
-                element.translate(dx, dy);
+                element.translate(dx as i16, dy as i16);
             }
         }
     }
